+++ R/nav-tabs.R
@@ -272,9 +272,9 @@ bs_tabset <- function(
     Negate(
       is.null
     ),
-    rlang::list2(
+    flatten_list_children(rlang::list2(
       ...
-    )
+    ))
   )
   ok <- vapply(
     panels,